+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Version 0.2.0
 
 - [BREAKING] change `inlineBandit` and `BoxedBandit` to type aliases of the now exported `Bandit`.
+- [ADDED] added policies and a policy paramter to `DoubleCollect`, determining state invalidation behaviour under contention.
 
 # Version 0.1.1